feat: 페이지 별 layout 적용 완료
