+++ src/app/(afterAuth)/mypage/DivWrapper/sections/ActionButtonSection.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import React from "react";
-import { Button } from "../../../../components/ui/button";
+import { Button } from "../../../../../components/ui/button";
 import {
   Bike,
   BookOpen,
