+++ src/app/(afterAuth)/mypage/DivWrapper/sections/DeleteMessage.tsx
