+++ src/app/(afterAuth)/mypage/DivWrapper/sections/FormContainerSection.tsx
@@ -1,8 +1,8 @@
 // FormContainerSection.tsx
 "use client";
 import React, { JSX } from "react";
-import { Button } from "../../../../components/ui/button";
-import { Card, CardContent } from "../../../../components/ui/card";
+import { Button } from "../../../../../components/ui/button";
+import { Card, CardContent } from "../../../../../components/ui/card";
 import { DeleteMessage } from "./DeleteMessage";
 
 export const FormContainerSection = (): JSX.Element => {
