+++ src/app/(afterAuth)/mypage/DivWrapper/sections/MyPost.tsx
@@ -1,13 +1,13 @@
 "use client";
 
 import React, { JSX } from "react";
-import { Avatar, AvatarFallback, AvatarImage } from "../../../../components/ui/avatar";
-import { Button } from "../../../../components/ui/button";
-import { Card, CardContent } from "../../../../components/ui/card";
-import { Tabs, TabsContent, TabsList, TabsTrigger } from "../../../../components/ui/tabs";
+import { Avatar, AvatarFallback, AvatarImage } from "../../../../../components/ui/avatar";
+import { Button } from "../../../../../components/ui/button";
+import { Card, CardContent } from "../../../../../components/ui/card";
+import { Tabs, TabsContent, TabsList, TabsTrigger } from "../../../../../components/ui/tabs";
 
 import PostCard from "@/components/common/PostCard";
-import { dummyPosts } from "@/app/posts/postData";
+import { dummyPosts } from "@/app/(afterAuth)/posts/postData";
 
 // 게시글 데이터
 const posts = [
