+++ src/app/(afterAuth)/mypage/DivWrapper/sections/ReadOnlyView.tsx
@@ -1,9 +1,9 @@
 "use client";
 import React from "react";
-import { Avatar, AvatarFallback, AvatarImage } from "../../../../components/ui/avatar";
-import { Badge } from "../../../../components/ui/badge";
-import { Button } from "../../../../components/ui/button";
-import { Card, CardContent } from "../../../../components/ui/card";
+import { Avatar, AvatarFallback, AvatarImage } from "../../../../../components/ui/avatar";
+import { Badge } from "../../../../../components/ui/badge";
+import { Button } from "../../../../../components/ui/button";
+import { Card, CardContent } from "../../../../../components/ui/card";
 
 // Data for category badges
 const categories = ["스포츠", "일상", "자기계발"];
