+++ src/app/(afterAuth)/mypage/DivWrapper/sections/Sidebar.tsx
@@ -1,6 +1,6 @@
 "use client";
 import React, { JSX } from "react";
-import { Card, CardContent } from "../../../../components/ui/card";
+import { Card, CardContent } from "../../../../../components/ui/card";
 
 type Props = {
   activeMenu: string;
