+++ src/app/(afterAuth)/mypage/DivWrapper/sections/UserInfoGroupSection.tsx
@@ -1,7 +1,7 @@
 "use client";
 import React, { JSX } from "react";
-import { Card, CardContent } from "../../../../components/ui/card";
-import { Switch } from "../../../../components/ui/switch";
+import { Card, CardContent } from "../../../../../components/ui/card";
+import { Switch } from "../../../../../components/ui/switch";
 
 export const UserInfoGroupSection = (): JSX.Element => {
   return (
