+++ src/app/(afterAuth)/participation/page.tsx
