+++ src/app/(afterAuth)/postDetail/page.tsx
