+++ src/app/(afterAuth)/posts/page.tsx
