+++ src/app/(afterAuth)/posts/postData.tsx
