+++ src/app/(afterAuth)/posts/write/FileInput.tsx
