+++ src/components/common/Navigation.tsx
@@ -12,7 +12,6 @@ import {
   DropdownMenu,
   DropdownMenuContent,
   DropdownMenuTrigger,
-  DropdownMenuPortal,
   DropdownMenuItem,
   DropdownMenuSeparator
 } from "@/components/ui/dropdown-menu";
@@ -195,7 +194,9 @@ export function Navigation() {
 
           <NavigationMenu>
             <NavigationMenuList>
-              {navItems.map((item) => (
+              {navItems
+              .filter(item => isLoggedIn || (item.name !== "마이페이지" && item.name!== "참여관리"))
+              .map((item) => (
                 <NavigationMenuItem key={item.href}>
                   <Link
                     href={item.href}
@@ -226,7 +227,7 @@ export function Navigation() {
                   <Bell className="h-6 w-6 text-foreground stroke-1.5" />
                 </button>
               </DropdownMenuTrigger>
-              <DropdownMenuPortal>
+
               <DropdownMenuContent
                 align="end"
                 sideOffset={4}
@@ -255,7 +256,6 @@ export function Navigation() {
                               </time>
                             </Link>
                         </DropdownMenuItem>
-                        
                       );
                     })}
                   </div>
@@ -268,7 +268,6 @@ export function Navigation() {
                       </Link>
                     </div>
                   </DropdownMenuContent>
-                  </DropdownMenuPortal>
               </DropdownMenu>
 
               <DropdownMenu>
@@ -290,7 +289,6 @@ export function Navigation() {
                     )}
                 </Avatar>
               </DropdownMenuTrigger>
-              <DropdownMenuPortal>
               <DropdownMenuContent align="end" sideOffset={4} className="w-80 h-70 bg-white border-gray-22 text-popover-foreground p-2">
                   <div className="flex flex-col items-center px-4 py-3 gap-3">
                     <div className = "text-title-md text-black self-start mb-2">나의 정보</div>
@@ -332,7 +330,6 @@ export function Navigation() {
                     <LogOut className="w-4 h-4 text-title-md" /> 로그아웃
                   </DropdownMenuItem>
                 </DropdownMenuContent>
-                </DropdownMenuPortal>
               </DropdownMenu>
             </>
           ) : (
